added authMiddleware, user id to all contacts methods

+++ model/contacts/removeContactById.js
@@ -1,7 +1,7 @@
 const { Contact } = require('../../db/contactModel')
 
-const removeContactById = async (id) => {
-  return await Contact.findByIdAndRemove({ _id: id })
+const removeContactById = async (id, _id) => {
+  return await Contact.findByIdAndRemove({ _id: id, owner: _id })
 }
 
 module.exports = removeContactById
